tests/wpt/web-platform-tests/2dcontext/the-canvas-state/2d.state.saverestore.textAlign.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: 2d.state.saverestore.textAlign</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/2dcontext/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.state.saverestore.textAlign</h1>
<p class="desc">save()/restore() works for textAlign</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("save()/restore() works for textAlign");
_addTest(function(canvas, ctx) {

// Test that restore() undoes any modifications
var old = ctx.textAlign;
ctx.save();
ctx.textAlign = "center";
ctx.restore();
_assertSame(ctx.textAlign, old, "ctx.textAlign", "old");

// Also test that save() doesn't modify the values
ctx.textAlign = "center";
old = ctx.textAlign;
    // we're not interested in failures caused by get(set(x)) != x (e.g.
    // from rounding), so compare against 'old' instead of against "center"
ctx.save();
_assertSame(ctx.textAlign, old, "ctx.textAlign", "old");
ctx.restore();


});
</script>
